+++ CHANGELOG.md
@@ -1,8 +1,11 @@
+# Head
+
+- Fixed: `selector-type-no-unknown` and `selector-type-case` rules now ignore placeholder selectors.
+
 # 6.4.2
 
 - Fixed: `selector-pseudo-class-case`, `selector-pseudo-class-no-unknown`, `selector-pseudo-element-case`, `selector-pseudo-element-no-unknown` rules now ignore SCSS variable interpolation.
-- Fixed: `selector-type-no-unknown` now ignores nested selectors, keyframe selectors, and placeholder selectors.
-- Fixed: `selector-type-case` now ignores placeholder selectors.
+- Fixed: `selector-type-no-unknown` now ignores nested selectors and keyframe selectors.
 
 # 6.4.1